% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{add_entities}
\alias{add_entities}
\title{Wrapper for \link[rocrateR]{add_entity}}
\usage{
add_entities(rocrate, entity, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{entity}{List with entity objects.}

\item{overwrite}{Boolean flag to indicate if the entity (if found in the
given RO-Crate) should be overwritten.}

\item{quiet}{Boolean flag to indicate if status messages should be hidden
(default: \code{FALSE}).}
}
\value{
Updated RO-Crate with the new entities.
}
\description{
Wrapper for \link[rocrateR]{add_entity}, can be use to add multiple entities.
}
