% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-entity.R
\name{.find_type_index}
\alias{.find_type_index}
\title{Find \verb{@type} index in RO-Crate}
\usage{
.find_type_index(rocrate, type)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}
}
\value{
Boolean vector with index for entity(ies) with \verb{@type}.
}
\description{
Find \verb{@type} index in RO-Crate. Useful to retrieve entities with a particular
type in the RO-Crate.
}
\keyword{internal}
