% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-files.R
\name{.get_zip_root}
\alias{.get_zip_root}
\title{Get root directory of zip file}
\usage{
.get_zip_root(zip_filename)
}
\arguments{
\item{zip_filename}{String with path to valid Zip file}
}
\value{
String with root directory for the given Zip file
}
\description{
Get root directory of zip file
}
\keyword{internal}
