% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.entity}
\alias{print.entity}
\title{Print RO-Crate entity}
\usage{
\method{print}{entity}(x, ...)
}
\arguments{
\item{x}{RO-Crate entity object, see \link{entity}.}

\item{...}{Optional arguments, not used.}
}
\value{
Invisibly the input RO-Crate entity, \code{x}.
}
\description{
Print RO-Crate entity, S3 method for class 'entity'.
}
\examples{
rocrateR::rocrate() |>
  rocrateR::get_entity("./")
}
