% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{remove_entities}
\alias{remove_entities}
\title{Wrapper for \link[rocrateR]{remove_entity}}
\usage{
remove_entities(rocrate, entity)
}
\arguments{
\item{rocrate}{RO-Crate object, see \link{rocrate}.}

\item{entity}{Entity object (list) that contains at least the following
components: \verb{@id} and \verb{@type}. Or, scalar value with entity \verb{@id}.}
}
\value{
Updated RO-Crate.
}
\description{
Wrapper for \link[rocrateR]{remove_entity}, can be use to remove multiple
entities.
}
